pyproject.toml
[tool.poetry]
name = "gen3"
homepage = "https://gen3.org/"
version = "3.1.0"
description = "Gen3 CLI and Python SDK"
authors = ["Center for Translational Data Science at the University of Chicago <support@datacommons.io>"]
license = "Apache-2.0"
packages = [
    { include = "gen3" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.6",
    "Topic :: Scientific/Engineering",
]

[tool.poetry.dependencies]
python = "^3.6"
requests = "*"
indexclient = ">=1.6.2"
drsclient = "<1.0.0"
aiohttp = "*"
backoff = "*"
click = "*"
pandas = "*"
pypfb = "<1.0.0"
rstr = "^2.2.6"
azure-storage-blob = "^12.6.0"

[tool.poetry.dev-dependencies]
pytest = ">=3.2.3"
cdisutilstest = { git = "https://github.com/uc-cdis/cdisutils-test.git", tag = "1.0.0" }
doiclient = { git = "https://github.com/uc-cdis/doiclient.git", tag = "1.0.0" }
dosclient = { git = "https://github.com/uc-cdis/dosclient.git", tag = "1.0.0" }
hsclient = { git = "https://github.com/uc-cdis/hsclient.git", tag = "1.0.0" }
indexd = { git = "https://github.com/uc-cdis/indexd.git", tag = "2.14.1" }

[tool.poetry.scripts]
gen3 = 'gen3.cli:main'

[tool.poetry.plugins."gen3.plugins"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
